pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "media-agent"
version = "0.3.0"
description = "AI agent for image generation with multi-provider LLM support and interactive CLI"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
keywords = ["image-generation", "z-image", "qwen", "llm", "langgraph", "agent", "diffusion", "ai", "cli"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    # Core ML
    "torch>=2.0.0",
    "diffusers @ git+https://github.com/huggingface/diffusers",
    "transformers>=4.40.0",
    "huggingface_hub>=0.20.0",
    "accelerate>=0.25.0",
    "bitsandbytes>=0.41.0",
    "einops>=0.7.0",
    "pillow>=10.0.0",
    # LangChain
    "langgraph>=0.2.0",
    "langchain-core>=0.3.0",
    # CLI
    "rich>=13.0.0",
    "prompt-toolkit>=3.0.40",
    # Config
    "pyyaml>=6.0",
    # HTTP (for API providers)
    "requests>=2.28.0",
]

[project.optional-dependencies]
# LLM providers
openai = ["openai>=1.0.0"]
anthropic = ["anthropic>=0.25.0"]
ollama = ["ollama>=0.1.0"]
all-llm = [
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "ollama>=0.1.0",
]
# Image preview
preview = ["rich-pixels>=3.0.0"]
# Development
dev = [
    "pytest>=7.0.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
]
# All optional dependencies
all = [
    "openai>=1.0.0",
    "anthropic>=0.25.0",
    "ollama>=0.1.0",
    "rich-pixels>=3.0.0",
]

[project.scripts]
media-agent = "media_agent.__main__:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[project.urls]
Homepage = "https://github.com/slahiri/media-agent"
Repository = "https://github.com/slahiri/media-agent"
Changelog = "https://github.com/slahiri/media-agent/blob/main/CHANGELOG.md"

[tool.setuptools.packages.find]
where = ["."]
include = ["media_agent*"]

[tool.setuptools.package-data]
media_agent = ["../config.yaml"]
